made changes in learn-more page and added TabContainer component
